requirements.txt
Django
djangorestframework
psycopg2
South
django-cors-headers
bpython
django-extensions
Werkzeug
